# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "{{.Description}}",
        "title": "{{.Title}}",
        "contact": {},
        "license": {},
        "version": "{{.Version}}"
    },
    "host": "{{.Host}}",
    "basePath": "{{.BasePath}}",
    "paths": {
...
        "/merchant/settings/payments": {
            "get": {
                "description": "获取承兑商的收款账户信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "承兑商APP API"
                ],
                "summary": "获取承兑商的收款账户信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户id",
                        "name": "uid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.GetPaymentsRet"
                        }
                    }
                }
            },
            "put": {
                "description": "修改承兑商的收款账户信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "承兑商APP API"
                ],
                "summary": "修改承兑商的收款账户信息",
                "parameters": [
                    {
                        "description": "输入参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.SetPaymentArg"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.SetPaymentRet"
                        }
                    }
                }
            },
            "post": {
                "description": "增加承兑商的收款账户信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "承兑商APP API"
                ],
                "summary": "增加承兑商的收款账户信息",
                "parameters": [
                    {
                        "description": "输入参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.AddPaymentArg"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.AddPaymentRet"
                        }
                    }
                }
            },
            "delete": {
                "description": "删除承兑商的收款账户信息。需要先确定是否有正在进行的订单，如果有不允许删除。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "承兑商APP API"
                ],
                "summary": "删除承兑商的收款账户信息",
                "parameters": [
                    {
                        "description": "输入参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.DeletePaymentArg"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.DeletePaymentRet"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "response.AddPaymentArg": {
            "type": "object",
            "properties": {
                "bank": {
                    "description": "所属银行",
                    "type": "string",
                    "example": "工商银行"
                },
                "bank_account": {
                    "description": "银行卡账号",
                    "type": "string",
                    "example": "88888888"
                },
                "bank_branch": {
                    "description": "所属银行分行",
                    "type": "string",
                    "example": "工商银行日本分行"
                },
                "gr_code": {
                    "description": "二维码信息",
                    "type": "string",
                    "example": "xxxx"
                },
                "name": {
                    "description": "账户名",
                    "type": "string",
                    "example": "sky"
                },
                "pay_type": {
                    "description": "支付类型 1:银行卡 2:微信 3:支付宝",
                    "type": "integer"
                },
                "uid": {
                    "description": "承兑商用户id",
                    "type": "integer"
                }
            }
        },
        "response.AddPaymentRet": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "entity": {
                    "type": "object",
                    "properties": {
                        "uid": {
                            "description": "用户id",
                            "type": "integer"
                        }
                    }
                },
                "err_code": {
                    "description": "err_code仅在失败时设置",
                    "type": "integer"
                },
                "err_msg": {
                    "description": "err_msg仅在失败时设置",
                    "type": "string",
                    "example": "由于xx原因，导致操作失败"
                },
                "status": {
                    "description": "status可以为success或者fail",
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "response.DeletePaymentArg": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "收款账号信息主键",
                    "type": "integer"
                },
                "uid": {
                    "description": "承兑商用户id",
                    "type": "integer"
                }
            }
        },
        "response.DeletePaymentRet": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "entity": {
                    "type": "object",
                    "properties": {
                        "uid": {
                            "description": "用户id",
                            "type": "integer"
                        }
                    }
                },
                "err_code": {
                    "description": "err_code仅在失败时设置",
                    "type": "integer"
                },
                "err_msg": {
                    "description": "err_msg仅在失败时设置",
                    "type": "string",
                    "example": "由于xx原因，导致操作失败"
                },
                "status": {
                    "description": "status可以为success或者fail",
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "response.GetPaymentsRet": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "entity": {
                    "type": "object",
                    "properties": {
                        "data": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/response.Payment"
                            }
                        },
                        "uid": {
                            "description": "用户id",
                            "type": "integer"
                        }
                    }
                },
                "err_code": {
                    "description": "err_code仅在失败时设置",
                    "type": "integer"
                },
                "err_msg": {
                    "description": "err_msg仅在失败时设置",
                    "type": "string",
                    "example": "由于xx原因，导致操作失败"
                },
                "status": {
                    "description": "status可以为success或者fail",
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "response.Payment": {
            "type": "object",
            "properties": {
                "bank": {
                    "description": "所属银行",
                    "type": "string",
                    "example": "工商银行"
                },
                "bank_account": {
                    "description": "银行卡账号",
                    "type": "string",
                    "example": "88888888"
                },
                "bank_branch": {
                    "description": "所属银行分行",
                    "type": "string",
                    "example": "工商银行日本分行"
                },
                "gr_code": {
                    "description": "二维码信息",
                    "type": "string",
                    "example": "xxxx"
                },
                "id": {
                    "description": "主键，无实际意义",
                    "type": "integer"
                },
                "name": {
                    "description": "账户名",
                    "type": "string",
                    "example": "sky"
                },
                "pay_type": {
                    "description": "支付类型 1:银行卡 2:微信 3:支付宝",
                    "type": "integer"
                }
            }
        },
...
        "response.SetPaymentArg": {
            "type": "object",
            "properties": {
                "bank": {
                    "description": "所属银行",
                    "type": "string",
                    "example": "工商银行"
                },
                "bank_account": {
                    "description": "银行卡账号",
                    "type": "string",
                    "example": "88888888"
                },
                "bank_branch": {
                    "description": "所属银行分行",
                    "type": "string",
                    "example": "工商银行日本分行"
                },
                "gr_code": {
                    "description": "二维码信息",
                    "type": "string",
                    "example": "xxxx"
                },
                "id": {
                    "description": "收款账号信息主键",
                    "type": "integer"
                },
                "name": {
                    "description": "账户名",
                    "type": "string",
                    "example": "sky"
                },
                "pay_type": {
                    "description": "支付类型 1:银行卡 2:微信 3:支付宝",
                    "type": "integer"
                }
            }
        },
        "response.SetPaymentRet": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "entity": {
                    "type": "object",
                    "properties": {
                        "uid": {
                            "description": "用户id",
                            "type": "integer"
                        }
                    }
                },
                "err_code": {
                    "description": "err_code仅在失败时设置",
                    "type": "integer"
                },
                "err_msg": {
                    "description": "err_msg仅在失败时设置",
                    "type": "string",
                    "example": "由于xx原因，导致操作失败"
                },
                "status": {
                    "description": "status可以为success或者fail",
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
    }
}